Gevent => Multiprocessing, DBM => Redis
Protocol parsing use a lot of processing time, and that is not an
ideal use case for gevent. Muitiprocessing make use of parallelism
to help with the processing. (Ideally some of the parsing should be
rewritten into Cython to reduce many Python lookup overhead).

Since each proxied connection take one process, each process is
effectively limited to using two threads. In this case, threading
will have a superior performance comparied to gevent. Not to mention
the incompatibility between multiprocessing and gevent.

Rcon management is propagated to a manager process from multiprocessing,
and each proxied connection will use one thread in the manager process.
Implementations of DBM does not seem to be all multiprocess-safe, and
switching to Redis solve this issue.