More work on item non-rendering bug
It seems that the packets spawn object and entity metadata must be
received by the client in a single recv. We use flush to workaround
this. However, flush is currently called by recv of both same direction
and opposite direction, and that creates a race condition where the
opposite direction flush out all the data when it shouldn't.

Also, since one recv call may not recv all the data because of the
circular buffer, we use select with 0 timeout to make sure that all
data are received.